Refactoring has begun.
Added comments up to line #156.

Still many functions to rearrange and document.

Several global functions in this project, however given that it was first coded some time ago and the hardware upon which it was designed has since been disassembled, it is not worth the time for me to refactor it to conform to better object oriented principles.